added transection to delete classroom

+++ src/app/modules/classroom/classroom.service.ts
@@ -16,17 +16,18 @@ const createClassroomIntoDB = async (payload: IClassroom, user: JwtPayload) => {
   if (isClassroomExists) {
     throw new Error("Classroom already exist!!!");
   }
-  const existingClasses = await Classroom.find({faculty:user._id}).select("classDays startTime endTime");
+  const existingClasses = await Classroom.find({ faculty: user._id }).select(
+    "classDays startTime endTime"
+  );
   const newSchedule = {
-    classDays:payload.classDays,
-    startTime:payload.startTime,
-    endTime:payload.endTime
-  }
-  const hasConflict = hasClassConflicts(newSchedule,existingClasses);
-  if(hasConflict){
+    classDays: payload.classDays,
+    startTime: payload.startTime,
+    endTime: payload.endTime,
+  };
+  const hasConflict = hasClassConflicts(newSchedule, existingClasses);
+  if (hasConflict) {
     throw new Error("Classroom time slot is already taken");
   }
-
 
   payload.faculty = user?._id;
   while (true) {
@@ -36,8 +37,12 @@ const createClassroomIntoDB = async (payload: IClassroom, user: JwtPayload) => {
     if (!isJoiningCodeExists) break;
   }
   const result = await Classroom.create(payload);
-  if(result){
-    updateClassCountSchedule(result._id.toString(), result.endTime, result.classDays)
+  if (result) {
+    updateClassCountSchedule(
+      result._id.toString(),
+      result.endTime,
+      result.classDays
+    );
   }
   return result;
 };
@@ -63,8 +68,26 @@ const deleteClassroomFromDB = async (classroomId: string, user: JwtPayload) => {
   if (classroom.faculty.toString() !== user._id) {
     throw new Error("You are unauthorized to delete the Classroom");
   }
-  const result = await Classroom.findByIdAndDelete(classroomId);
-  return result;
+  const session = await mongoose.startSession()
+  try {
+    session.startTransaction()
+    const deletedAttendance = await Attendance.deleteMany({classroom:classroomId})
+    if(!deletedAttendance){
+      throw new Error("Something went wrong")
+    }
+    const result = await Classroom.findByIdAndDelete(classroomId);
+    if(!result){
+      throw new Error("Something went wrong")
+    }
+    await session.commitTransaction()
+    await session.endSession()
+    return result;
+
+  } catch (error) {
+    await session.abortTransaction()
+    await session.endSession()
+    throw new Error("Something went wrong")
+  }
 };
 
 const joinClassroom = async (joiningCode: string, user: JwtPayload) => {
@@ -81,44 +104,44 @@ const joinClassroom = async (joiningCode: string, user: JwtPayload) => {
   }
   const myClassrooms = await Attendance.aggregate([
     {
-        $match: {
-            student: new mongoose.Types.ObjectId(user._id)
-        }
+      $match: {
+        student: new mongoose.Types.ObjectId(user._id),
+      },
     },
     {
-        $lookup: {
-            from: "classrooms",
-            localField: "classroom",
-            foreignField: "_id",
-            as: "classroomDetails"
-        }
+      $lookup: {
+        from: "classrooms",
+        localField: "classroom",
+        foreignField: "_id",
+        as: "classroomDetails",
+      },
     },
     {
-        $unwind: "$classroomDetails"
+      $unwind: "$classroomDetails",
     },
     {
-        $replaceRoot: { 
-            newRoot: "$classroomDetails" 
-        }
+      $replaceRoot: {
+        newRoot: "$classroomDetails",
+      },
     },
     {
-      $project:{
-        _id:0,
-        classDays:1,
-        startTime:1,
-        endTime:1
-      }
-    }
-])
-const newSchedule = {
-  classDays:isClassroomExists.classDays,
-  startTime:isClassroomExists.startTime,
-  endTime:isClassroomExists.endTime
-}
-const hasClassConflict = hasClassConflicts(newSchedule,myClassrooms);
-if(hasClassConflict){
-  throw new Error("Classroom time slot is already taken");
-}
+      $project: {
+        _id: 0,
+        classDays: 1,
+        startTime: 1,
+        endTime: 1,
+      },
+    },
+  ]);
+  const newSchedule = {
+    classDays: isClassroomExists.classDays,
+    startTime: isClassroomExists.startTime,
+    endTime: isClassroomExists.endTime,
+  };
+  const hasClassConflict = hasClassConflicts(newSchedule, myClassrooms);
+  if (hasClassConflict) {
+    throw new Error("Classroom time slot is already taken");
+  }
   const attendance = {
     classroom: isClassroomExists._id,
     student: user._id,
@@ -139,38 +162,50 @@ const leaveClassroom = async (classroomId: string, user: JwtPayload) => {
   return result;
 };
 
-const getMyClassrooms = async(user:JwtPayload)=>{
-    if(user.role === userRoles.STUDENT){
-        const result = await Attendance.aggregate([
-            {
-                $match: {
-                    student: new mongoose.Types.ObjectId(user._id)
-                }
-            },
-            {
-                $lookup: {
-                    from: "classrooms",
-                    localField: "classroom",
-                    foreignField: "_id",
-                    as: "classroomDetails"
-                }
-            },
-            {
-                $unwind: "$classroomDetails"
-            },
-            {
-                $replaceRoot: { 
-                    newRoot: "$classroomDetails" 
-                }
-            }
-        ])
-        return result
-    }else if(user.role === userRoles.FACULTY){
-      const result = await Classroom.find({faculty:user._id})
-      return result
-    }
-    return null
-}
+const getMyClassrooms = async (user: JwtPayload) => {
+  if (user.role === userRoles.STUDENT) {
+    const result = await Attendance.aggregate([
+      {
+        $match: {
+          student: new mongoose.Types.ObjectId(user._id),
+        },
+      },
+      {
+        $lookup: {
+          from: "classrooms",
+          localField: "classroom",
+          foreignField: "_id",
+          as: "classroomDetails",
+        },
+      },
+      {
+        $unwind: "$classroomDetails",
+      },
+      {
+        $replaceRoot: {
+          newRoot: "$classroomDetails",
+        },
+      },
+      {
+        $lookup: {
+          from: "users",
+          localField: "faculty",
+          foreignField: "_id",
+          as: "faculty",
+        },
+      },
+
+      {
+        $unwind: "$faculty",
+      },
+    ]);
+    return result;
+  } else if (user.role === userRoles.FACULTY) {
+    const result = await Classroom.find({ faculty: user._id });
+    return result;
+  }
+  return null;
+};
 
 export const ClassroomService = {
   createClassroomIntoDB,
@@ -179,5 +214,5 @@ export const ClassroomService = {
   deleteClassroomFromDB,
   joinClassroom,
   leaveClassroom,
-  getMyClassrooms
+  getMyClassrooms,
 };
